Text file input uses Named Entity Recognition and Input error checking
Changed text file input to use named entity recognition from nltk. Can
now recognize novel names. It overgeneralized words as names though, so
I made it check against a list of english words (wordsEn.txt) before
accepting it as a name. This makes some names that are also words not
count, but it’s one or the other and I think this is preferable.

Also implemented some input catches so that if the user leaves some
fields blank it doesn’t throw errors.

This commit also includes HarryPotter.txt, which is Chapter 7 of Book 1
of Harry Potter. This is to be used for a demo.